Set up gameFlags to be properly saved/loaded.
