+++ src/types/saveArrays.ts
@@ -1,0 +1,16 @@
+export interface SaveUpgradeArray {
+    key: number,
+    unlocked: boolean,
+    bought: boolean,
+}
+
+export interface SaveKillsArray {
+    key: string,
+    kills: number,
+}
+
+export interface SavedGameFlags {
+    key: number,
+    description: string,
+    state: boolean
+}
