--- src/types/saveKillsArray.ts
@@ -1,4 +1,0 @@
-export interface SaveKillsArray {
-    key: string,
-    kills: number,
-}
