--- src/types/saveUpgradeArray.ts
@@ -1,5 +1,0 @@
-export interface SaveUpgradeArray {
-    key: number,
-    unlocked: boolean,
-    bought: boolean,
-}
